make eslint happy
this should make the test on travis-ci.org pass. Acually, this is a work-around
rather than a fix, because it would have been quite cumbersome to add
"use strict" statements everywhere. This would lead to merge conflicts, and the
benefit is not that high. Once switched to a module system, this check can be
enabled, again.